+++ update-server/otupdate/common/name_management.py
